requirements.txt
feedparser
requests
python-dotenv
supabase
